bug fixes and cleanups

+++ app/src/main/java/javax/vecmath/Matrix3f.java
@@ -1949,15 +1949,15 @@ public class Matrix3f implements java.io.Serializable, Cloneable {
      */  
     public int hashCode() {
 	long bits = 1L;
-	bits = 31L * bits + (long)VecMathUtil.floatToIntBits(m00);
-	bits = 31L * bits + (long)VecMathUtil.floatToIntBits(m01);
-	bits = 31L * bits + (long)VecMathUtil.floatToIntBits(m02);
-	bits = 31L * bits + (long)VecMathUtil.floatToIntBits(m10);
-	bits = 31L * bits + (long)VecMathUtil.floatToIntBits(m11);
-	bits = 31L * bits + (long)VecMathUtil.floatToIntBits(m12);
-	bits = 31L * bits + (long)VecMathUtil.floatToIntBits(m20);
-	bits = 31L * bits + (long)VecMathUtil.floatToIntBits(m21);
-	bits = 31L * bits + (long)VecMathUtil.floatToIntBits(m22);
+	bits = 31L * bits + VecMathUtil.floatToIntBits(m00);
+	bits = 31L * bits + VecMathUtil.floatToIntBits(m01);
+	bits = 31L * bits + VecMathUtil.floatToIntBits(m02);
+	bits = 31L * bits + VecMathUtil.floatToIntBits(m10);
+	bits = 31L * bits + VecMathUtil.floatToIntBits(m11);
+	bits = 31L * bits + VecMathUtil.floatToIntBits(m12);
+	bits = 31L * bits + VecMathUtil.floatToIntBits(m20);
+	bits = 31L * bits + VecMathUtil.floatToIntBits(m21);
+	bits = 31L * bits + VecMathUtil.floatToIntBits(m22);
 	return (int) (bits ^ (bits >> 32));
     }
